scripts/requirements.txt
numpy>=1.24
Pillow>=10.0
kagglehub>=0.3
